Move fixture store and snapshot store to different files
