+++ src/storage/txn/store/mod.rs
@@ -1,14 +1,13 @@
-// Copyright 2016 TiKV Project Authors. Licensed under Apache-2.0.
+// Copyright 2019 TiKV Project Authors. Licensed under Apache-2.0.
 
-use std::collections::{btree_map, BTreeMap};
-use std::sync::Arc;
+mod fixture_store;
+mod snapshot_store;
 
-use kvproto::kvrpcpb::IsolationLevel;
+pub use self::fixture_store::{FixtureStore, FixtureStoreScanner};
+pub use self::snapshot_store::SnapshotStore;
 
-use crate::storage::metrics::*;
-use crate::storage::mvcc::{Error as MvccError, MvccReader};
-use crate::storage::mvcc::{Scanner as MvccScanner, ScannerBuilder};
-use crate::storage::{Key, KvPair, Snapshot, Statistics, Value};
+use crate::storage::mvcc::Error as MvccError;
+use crate::storage::{Key, KvPair, Statistics, Value};
 
 use super::{Error, Result};
 
@@ -49,231 +48,6 @@ pub trait Scanner: Send {
     }
 
     fn take_statistics(&mut self) -> Statistics;
-}
-
-pub struct SnapshotStore<S: Snapshot> {
-    snapshot: S,
-    start_ts: u64,
-    isolation_level: IsolationLevel,
-    fill_cache: bool,
-}
-
-impl<S: Snapshot> Store for SnapshotStore<S> {
-    type Scanner = MvccScanner<S>;
-
-    #[inline]
-    fn get(&self, key: &Key, statistics: &mut Statistics) -> Result<Option<Value>> {
-        let mut reader = MvccReader::new(
-            self.snapshot.clone(),
-            None,
-            self.fill_cache,
-            None,
-            None,
-            self.isolation_level,
-        );
-        let v = reader.get(key, self.start_ts)?;
-        statistics.add(reader.get_statistics());
-        Ok(v)
-    }
-
-    #[inline]
-    fn batch_get(&self, keys: &[Key], statistics: &mut Statistics) -> Vec<Result<Option<Value>>> {
-        // TODO: sort the keys and use ScanMode::Forward
-        let mut reader = MvccReader::new(
-            self.snapshot.clone(),
-            None,
-            self.fill_cache,
-            None,
-            None,
-            self.isolation_level,
-        );
-        let mut results = Vec::with_capacity(keys.len());
-        for k in keys {
-            results.push(reader.get(k, self.start_ts).map_err(Error::from));
-        }
-        statistics.add(reader.get_statistics());
-        results
-    }
-
-    #[inline]
-    fn scanner(
-        &self,
-        desc: bool,
-        key_only: bool,
-        lower_bound: Option<Key>,
-        upper_bound: Option<Key>,
-    ) -> Result<MvccScanner<S>> {
-        // Check request bounds with physical bound
-        self.verify_range(&lower_bound, &upper_bound)?;
-        let scanner = ScannerBuilder::new(self.snapshot.clone(), self.start_ts, desc)
-            .range(lower_bound, upper_bound)
-            .omit_value(key_only)
-            .fill_cache(self.fill_cache)
-            .isolation_level(self.isolation_level)
-            .build()?;
-
-        Ok(scanner)
-    }
-}
-
-impl<S: Snapshot> SnapshotStore<S> {
-    pub fn new(
-        snapshot: S,
-        start_ts: u64,
-        isolation_level: IsolationLevel,
-        fill_cache: bool,
-    ) -> Self {
-        SnapshotStore {
-            snapshot,
-            start_ts,
-            isolation_level,
-            fill_cache,
-        }
-    }
-
-    fn verify_range(&self, lower_bound: &Option<Key>, upper_bound: &Option<Key>) -> Result<()> {
-        if let Some(ref l) = lower_bound {
-            if let Some(b) = self.snapshot.lower_bound() {
-                if !b.is_empty() && l.as_encoded().as_slice() < b {
-                    REQUEST_EXCEED_BOUND.inc();
-                    return Err(Error::InvalidReqRange {
-                        start: Some(l.as_encoded().clone()),
-                        end: upper_bound.as_ref().map(|ref b| b.as_encoded().clone()),
-                        lower_bound: Some(b.to_vec()),
-                        upper_bound: self.snapshot.upper_bound().map(|b| b.to_vec()),
-                    });
-                }
-            }
-        }
-        if let Some(ref u) = upper_bound {
-            if let Some(b) = self.snapshot.upper_bound() {
-                if !b.is_empty() && (u.as_encoded().as_slice() > b || u.as_encoded().is_empty()) {
-                    REQUEST_EXCEED_BOUND.inc();
-                    return Err(Error::InvalidReqRange {
-                        start: lower_bound.as_ref().map(|ref b| b.as_encoded().clone()),
-                        end: Some(u.as_encoded().clone()),
-                        lower_bound: self.snapshot.lower_bound().map(|b| b.to_vec()),
-                        upper_bound: Some(b.to_vec()),
-                    });
-                }
-            }
-        }
-
-        Ok(())
-    }
-}
-
-/// A Store that reads on fixtures.
-#[derive(Clone)]
-pub struct FixtureStore {
-    data: Arc<BTreeMap<Key, Result<Vec<u8>>>>,
-}
-
-impl FixtureStore {
-    pub fn new(data: BTreeMap<Key, Result<Vec<u8>>>) -> Self {
-        FixtureStore {
-            data: Arc::new(data),
-        }
-    }
-}
-
-impl Store for FixtureStore {
-    type Scanner = FixtureStoreScanner;
-
-    #[inline]
-    fn get(&self, key: &Key, _statistics: &mut Statistics) -> Result<Option<Vec<u8>>> {
-        let r = self.data.get(key);
-        match r {
-            None => Ok(None),
-            Some(Ok(v)) => Ok(Some(v.clone())),
-            Some(Err(e)) => Err(e.maybe_clone().unwrap()),
-        }
-    }
-
-    #[inline]
-    fn batch_get(&self, keys: &[Key], statistics: &mut Statistics) -> Vec<Result<Option<Vec<u8>>>> {
-        keys.iter().map(|key| self.get(key, statistics)).collect()
-    }
-
-    #[inline]
-    fn scanner(
-        &self,
-        desc: bool,
-        key_only: bool,
-        lower_bound: Option<Key>,
-        upper_bound: Option<Key>,
-    ) -> Result<FixtureStoreScanner> {
-        use std::ops::Bound;
-
-        let lower = lower_bound.as_ref().map_or(Bound::Unbounded, |v| {
-            if !desc {
-                Bound::Included(v)
-            } else {
-                Bound::Excluded(v)
-            }
-        });
-        let upper = upper_bound.as_ref().map_or(Bound::Unbounded, |v| {
-            if desc {
-                Bound::Included(v)
-            } else {
-                Bound::Excluded(v)
-            }
-        });
-
-        let range_data_ref = self.data.clone();
-        // Erase the lifetime to be 'static.
-        let range_unsafe = unsafe { std::mem::transmute(range_data_ref.range((lower, upper))) };
-
-        Ok(FixtureStoreScanner {
-            key_only,
-            desc,
-            _range_data_ref: range_data_ref,
-            range_unsafe,
-        })
-    }
-}
-
-/// A Scanner that scans on fixtures.
-// TODO: Replace it using GATs when available, to avoid `Arc` cost.
-#[allow(clippy::type_complexity)]
-pub struct FixtureStoreScanner {
-    key_only: bool,
-    desc: bool,
-    /// Accessing this field is unsafe, since it's lifetime is faked. This field is valid only if
-    /// `range_data_ref` is valid. When `range_data_ref` is destroyed, this field is no longer
-    /// valid. Thus we must *NEVER* leak the reference of this field to the outside.
-    range_unsafe: btree_map::Range<'static, Key, Result<Vec<u8>>>,
-    _range_data_ref: Arc<BTreeMap<Key, Result<Vec<u8>>>>,
-}
-
-impl Scanner for FixtureStoreScanner {
-    #[inline]
-    fn next(&mut self) -> Result<Option<(Key, Vec<u8>)>> {
-        let value = if !self.desc {
-            // During the call of this function, `range_data_ref` must be valid and we are only
-            // returning data clones to outside, so this access is safe, which does not violate
-            // the *real* life time restriction.
-            self.range_unsafe.next()
-        } else {
-            self.range_unsafe.next_back()
-        };
-        match value {
-            None => Ok(None),
-            Some((k, Ok(v))) => {
-                if !self.key_only {
-                    Ok(Some((k.clone(), v.clone())))
-                } else {
-                    Ok(Some((k.clone(), Vec::new())))
-                }
-            }
-            Some((_k, Err(e))) => Err(e.maybe_clone().unwrap()),
-        }
-    }
-
-    #[inline]
-    fn take_statistics(&mut self) -> Statistics {
-        Statistics::default()
-    }
 }
 
 #[cfg(test)]
